fixes for comments from audit

+++ contracts/interfaces/IIssuer.sol
@@ -93,5 +93,5 @@ interface IIssuer {
         address liquidator
     ) external returns (uint totalRedeemed, uint amountToLiquidate);
 
-    function setCurrentPeriodId(uint periodId) external;
+    function setCurrentPeriodId(uint128 periodId) external;
 }
